Cleaned up server.js
