+++ server.js
@@ -30,6 +30,7 @@ app.get('/users', passport.authenticate('bearer'), (req, res) => {
 })
 
 app.use(require("./routes/api_routes"));
+app.use(require("./routes/recipe_routes"));
 
 app.listen(PORT, () => {
     console.log(`🌎 ==> API server now on port ${PORT}!`);
